Changed chief_rate column type on users table

+++ database/migrations/2019_09_16_171258_add_chief_rate_on_users_table.php
@@ -14,7 +14,7 @@ class AddChiefRateOnUsersTable extends Migration
     public function up()
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->unsignedInteger('chief_rate')->default(0);
+            $table->decimal('chief_rate', 8, 2)->default(0);
         });
     }
